+++ lib/open_meteo.rb
@@ -1,2 +1,3 @@
-require "open_meteo/forecast"
+require "open_meteo/types"
 require "open_meteo/location"
+require "open_meteo/forecast"