+++ lib/open_meteo/location.rb
@@ -1,17 +1,14 @@
+require_relative "location_contract"
+
 module OpenMeteo
   ##
   # A location for a request to OpenMeteo.
-  class Location
-    attr_reader :latitude, :longitude
-
-    def initialize(longitude:, latitude:)
-      @longitude = longitude
-      @latitude = latitude
-    end
+  class Location < Dry::Struct
+    attribute :latitude, OpenMeteo::Types::Strict::Float
+    attribute :longitude, OpenMeteo::Types::Strict::Float
 
     def validate!
-      # FIXME: Placeholder for validation
-      true
+      LocationContract.validate!(to_hash)
     end
 
     def to_get_params