+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+- Use [`dry-validation`](https://github.com/dry-rb/dry-validation) to validate data classes
+- Use [`dry-types`](https://github.com/dry-rb/dry-types) and [`dry-struct`](https://github.com/dry-rb/dry-struct) for data classes
+
 ## [0.0.2] - 2023-11-03
 
 - Adjust Faraday version dependency